requirements.txt
mcp>=0.1.0
pydantic>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
typer>=0.9.0
requests>=2.31.0
python-dotenv>=1.0.0
anthropic>=0.34.0
fastapi>=0.110.0
uvicorn>=0.29.0